requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
tqdm
python-dotenv
google-api-python-client
selenium
webdriver-manager
beautifulsoup4
regex
